requirements.txt
django-filter
Django>=5.2,<5.3
wagtail>=7.1,<7.2
